Factor typed_array into a first class type
